refactor: proper separation from types and services
